* Moved dualsink.h and logger.h to logging folder in include/flashbackclient
