+++ src/managers/rulemanager.cpp
@@ -1,4 +1,4 @@
-#include <flashbackclient/logger.h>
+#include <flashbackclient/logging/logger.h>
 #include <flashbackclient/managers/rulemanager.h>
 #include <memory>
