+++ src/managers/settingmanager.cpp
@@ -1,6 +1,6 @@
 #include <flashbackclient/managers/settingmanager.h>
 
-#include <flashbackclient/logger.h>
+#include <flashbackclient/logging/logger.h>
 #include <yaml-cpp/yaml.h>
 
 namespace FlashBackClient
