+++ src/rule.cpp
@@ -1,7 +1,7 @@
 #include <flashbackclient/rule.h>
 
 #include <flashbackclient/condition.h>
-#include <flashbackclient/logger.h>
+#include <flashbackclient/logging/logger.h>
 #include <flashbackclient/target.h>
 #include <memory>
 #include <unordered_map>
