+++ src/signal_handler.cpp
@@ -1,7 +1,7 @@
 #include <flashbackclient/signal_handler.h>
 
 #include <flashbackclient/configs.h>
-#include <flashbackclient/logger.h>
+#include <flashbackclient/logging/logger.h>
 #include <flashbackclient/scheduler.h>
 #include <flashbackclient/service_locator.h>
