+++ src/target.cpp
@@ -1,6 +1,6 @@
 #include <flashbackclient/target.h>
 
-#include <flashbackclient/logger.h>
+#include <flashbackclient/logging/logger.h>
 #include <flashbackclient/rule.h>
 #include <flashbackclient/scheduler.h>
 #include <flashbackclient/service_locator.h>
